backend/requirements.txt
Flask==2.0.3
flask-cors==3.0.10

psycopg2-binary==2.9.3
SQLAlchemy[mypy]==1.4.35
Flask-SQLAlchemy==2.5.1
alembic==1.7.7
Flask-Migrate==3.1.0

flask-smorest==0.37.0
marshmallow==3.15.0

Pillow==9.1.0

pytest==7.1.1
pytest-cov==3.0.0
# pytest-flask-sqlalchemy==1.0.2
# Install package from Github, because of the bug: https://github.com/jeancochrane/pytest-flask-sqlalchemy/issues/50
# It's fixed: https://github.com/jeancochrane/pytest-flask-sqlalchemy/pull/51
# But new version is not released yet.
pytest-flask-sqlalchemy @ git+https://github.com/jeancochrane/pytest-flask-sqlalchemy.git@ea5c5ee935ed2d9cdc56061c243a6004d7c8fe7e
pytest-mock==3.7.0
